Fix bug with disappearing search results
When clicking the header to sort the table of results, a GET request is sent to the server.  In this instance GET requests are typically considered a page render without fetching any results from the database.  For example, when you first load the page no results are shown to you.  As such, clicking the header to change the ordering by perspective (asc, desc) is just a GET request, which renders the page without conducting a search.  Additionally, the query parameter sortdir is used for GET requests, however, POST methods use sortDir meaning  the user will get inconsistent ordered results; they might sort a column DESC and then set a filter attribute that will POST back to the server, fail to see the DESC property and fall back to ASC.